Removed `node.go` and moved its methods to more appropriate places
