+++ suffuse/idnode.go
@@ -93,3 +93,18 @@ func (x *IdNode) ReadAll(ctx context.Context) ([]byte, error) {
   }
   return nil, NotExist()
 }
+
+// HandleRead handles a read request assuming that data is the entire file content.
+// It adjusts the amount returned in resp according to req.Offset and req.Size.
+func HandleRead(req *f.ReadRequest, resp *f.ReadResponse, data []byte) {
+  if req.Offset >= int64(len(data)) {
+    data = nil
+  } else {
+    data = data[req.Offset:]
+  }
+  if len(data) > req.Size {
+    data = data[:req.Size]
+  }
+  n := copy(resp.Data[:req.Size], data)
+  resp.Data = resp.Data[:n]
+}
